requirements.txt
black==23.12.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
python-dotenv==1.0.0
regex==2023.12.25
requests==2.31.0
tiktoken==0.5.2
typing_extensions==4.9.0
unique-sdk==0.5.0
urllib3==2.1.0
Werkzeug==3.0.1
